Split workflow YAML behavior tests
