+++ test/Aevatar.Workflow.Host.Api.Tests/WorkflowDefinitionParserResourceLimitTests.cs
@@ -1,0 +1,55 @@
+using System.Text;
+using Aevatar.Workflow.Application.Abstractions.Runs;
+using Aevatar.Workflow.Core;
+using Aevatar.Workflow.Infrastructure.Runs;
+using FluentAssertions;
+
+namespace Aevatar.Workflow.Host.Api.Tests;
+
+public sealed class WorkflowDefinitionParserResourceLimitTests
+{
+    [Fact]
+    public async Task ParseWorkflowYamlAsync_ShouldClassifyResourceLimit()
+    {
+        var parser = new WorkflowDefinitionParser([new WorkflowCoreModulePack()]);
+
+        var result = await parser.ParseWorkflowYamlAsync(BuildNestedWorkflowYaml(childLinks: 31));
+
+        result.Succeeded.Should().BeFalse();
+        result.ErrorCode.Should().Be(WorkflowYamlParseErrorCode.ResourceLimit);
+        result.Error.Should().Contain("nesting depth");
+    }
+
+    [Fact]
+    public async Task ParseInlineWorkflowBundleAsync_ShouldPropagateResourceLimit()
+    {
+        var parser = new WorkflowDefinitionParser([new WorkflowCoreModulePack()]);
+
+        var result = await parser.ParseInlineWorkflowBundleAsync(
+            [new WorkflowChatInlineYamlDocument(string.Empty, BuildNestedWorkflowYaml(childLinks: 31))]);
+
+        result.Succeeded.Should().BeFalse();
+        result.ErrorCode.Should().Be(WorkflowYamlParseErrorCode.ResourceLimit);
+        result.Error.Should().Contain("nesting depth");
+    }
+
+    private static string BuildNestedWorkflowYaml(int childLinks)
+    {
+        var yaml = new StringBuilder()
+            .AppendLine("name: nested")
+            .AppendLine("roles: []")
+            .AppendLine("steps:");
+
+        for (var index = 0; index <= childLinks; index++)
+        {
+            var itemIndent = new string(' ', 2 + (index * 4));
+            var propertyIndent = new string(' ', 4 + (index * 4));
+            yaml.Append(itemIndent).Append("- id: step_").AppendLine(index.ToString());
+            yaml.Append(propertyIndent).AppendLine("type: assign");
+            if (index < childLinks)
+                yaml.Append(propertyIndent).AppendLine("children:");
+        }
+
+        return yaml.ToString();
+    }
+}
